+++ src-tauri/src/groups.rs
@@ -1,4 +1,6 @@
 use byteorder::{BigEndian, ReadBytesExt};
+use rdkafka::admin::{AdminClient, AdminOptions};
+use rdkafka::client::DefaultClientContext;
 use rdkafka::consumer::{CommitMode, Consumer, StreamConsumer};
 use rdkafka::{ClientConfig, Offset, TopicPartitionList};
 use serde::Serialize;
@@ -151,7 +153,7 @@ pub async fn get_groups_from_topic(
     Ok(groups_list)
 }
 
-pub async fn reset_offsets(
+pub async fn commit_latest_offsets(
     mut common_config: ClientConfig,
     group_name: String,
     topic_name: String,
@@ -190,6 +192,49 @@ pub async fn reset_offsets(
         .commit(&tpl, CommitMode::Sync)
         .map_err(|err| format!("Could not commit offsets: {}", err.to_string()))?;
 
+    Ok(())
+}
+
+pub async fn seek_earliest_offsets(
+    mut common_config: ClientConfig,
+    group_name: String,
+    topic_name: String,
+) -> Result<(), String> {
+    common_config.set("group.id", group_name);
+    let consumer: StreamConsumer = common_config.create().map_err(|err| {
+        format!(
+            "Could not create consumer to fetch offsets: {}",
+            err.to_string()
+        )
+    })?;
+
+    let mut tpl = TopicPartitionList::new();
+
+    let metadata = consumer
+        .fetch_metadata(Some(&topic_name), Duration::from_secs(30))
+        .map_err(|err| format!("Could not get metadata from cluster: {}", err.to_string()))?;
+
+    for partition in metadata.topics().get(0).unwrap().partitions() {
+        tpl.add_partition_offset(&topic_name, partition.id(), Offset::Offset(0))
+            .unwrap();
+    }
+
+    consumer
+        .commit(&tpl, CommitMode::Sync)
+        .map_err(|err| format!("Could not commit offsets: {}", err.to_string()))?;
+
+    Ok(())
+}
+
+pub async fn delete_group(
+    admin: &AdminClient<DefaultClientContext>,
+    group_name: String,
+) -> Result<(), String> {
+    let opts = AdminOptions::new().request_timeout(Some(Duration::from_secs(10)));
+    admin
+        .delete_groups(&[&group_name], &opts)
+        .await
+        .map_err(|err| format!("Error deleting consumer group: {}", err.to_string()))?;
     Ok(())
 }